+++ include/swift/Basic/Statistics.def
@@ -227,6 +227,7 @@ FRONTEND_STATISTIC(Sema, NumUnloadedLazyIterableDeclContexts)
 #include "swift/AST/AccessTypeIDZone.def"
 #include "swift/AST/NameLookupTypeIDZone.def"
 #include "swift/AST/TypeCheckerTypeIDZone.def"
+#include "swift/IDE/IDERequestIDZone.def"
 #undef SWIFT_TYPEID
 
 /// The next 10 statistics count 5 kinds of SIL entities present
