+++ include/swift/Subsystems.h
@@ -396,6 +396,11 @@ namespace swift {
   /// ASTContext.
   void registerTypeCheckerRequestFunctions(Evaluator &evaluator);
 
+  /// Register IDE-level request functions with the evaluator.
+  ///
+  /// The ASTContext will automatically call these upon construction.
+  void registerIDERequestFunctions(Evaluator &evaluator);
+
 } // end namespace swift
 
 #endif // SWIFT_SUBSYSTEMS_H
