+++ lib/IDE/SwiftSourceDocInfo.cpp
@@ -68,213 +68,8 @@ void XMLEscapingPrinter::printXML(StringRef Text) {
   OS << Text;
 }
 
-SourceManager &CursorInfoResolver::getSourceMgr() const
-{
-  return SrcFile.getASTContext().SourceMgr;
-}
-
-bool CursorInfoResolver::tryResolve(ValueDecl *D, TypeDecl *CtorTyRef,
-                                 ExtensionDecl *ExtTyRef, SourceLoc Loc,
-                                 bool IsRef, Type Ty) {
-  if (!D->hasName())
-    return false;
-
-  if (Loc != LocToResolve)
-    return false;
-
-  if (auto *VD = dyn_cast<VarDecl>(D)) {
-    // Handle references to the implicitly generated vars in case statements
-    // matching multiple patterns
-    if (VD->isImplicit()) {
-      if (auto * Parent = VD->getParentVarDecl()) {
-        D = Parent;
-        VD = Parent;
-      }
-    }
-  }
-  CursorInfo.setValueRef(D, CtorTyRef, ExtTyRef, IsRef, Ty, ContainerType);
-  return true;
-}
-
-bool CursorInfoResolver::tryResolve(ModuleEntity Mod, SourceLoc Loc) {
-  if (Loc == LocToResolve) {
-    CursorInfo.setModuleRef(Mod);
-    return true;
-  }
-  return false;
-}
-
-bool CursorInfoResolver::tryResolve(Stmt *St) {
-  if (auto *LST = dyn_cast<LabeledStmt>(St)) {
-    if (LST->getStartLoc() == LocToResolve) {
-      CursorInfo.setTrailingStmt(St);
-      return true;
-    }
-  }
-  if (auto *CS = dyn_cast<CaseStmt>(St)) {
-    if (CS->getStartLoc() == LocToResolve) {
-      CursorInfo.setTrailingStmt(St);
-      return true;
-    }
-  }
-  return false;
-}
-
-bool CursorInfoResolver::visitSubscriptReference(ValueDecl *D,
-                                                 CharSourceRange Range,
-                                                 ReferenceMetaData Data,
-                                                 bool IsOpenBracket) {
-  // We should treat both open and close brackets equally
-  return visitDeclReference(D, Range, nullptr, nullptr, Type(), Data);
-}
-
-ResolvedCursorInfo CursorInfoResolver::resolve(SourceLoc Loc) {
-  assert(Loc.isValid());
-  LocToResolve = Loc;
-  CursorInfo.Loc = Loc;
-  walk(SrcFile);
-  return CursorInfo;
-}
-
-bool CursorInfoResolver::walkToDeclPre(Decl *D, CharSourceRange Range) {
-  if (!rangeContainsLoc(D->getSourceRangeIncludingAttrs()))
-    return false;
-
-  if (isa<ExtensionDecl>(D))
-    return true;
-
-  if (auto *VD = dyn_cast<ValueDecl>(D))
-    return !tryResolve(VD, /*CtorTyRef=*/nullptr, /*ExtTyRef=*/nullptr,
-                       Range.getStart(), /*IsRef=*/false);
-
-  return true;
-}
-
-bool CursorInfoResolver::walkToDeclPost(Decl *D) {
-  if (isDone())
-    return false;
-  if (getSourceMgr().isBeforeInBuffer(LocToResolve, D->getStartLoc()))
-    return false;
-  return true;
-}
-
-bool CursorInfoResolver::walkToStmtPre(Stmt *S) {
-  // Getting the character range for the statement, to account for interpolation
-  // strings. The token range for the interpolation string is the whole string,
-  // with begin/end locations pointing at the beginning of the string, so if
-  // there is a token location inside the string, it will seem as if it is out
-  // of the source range, unless we convert to character range.
-
-  // FIXME: Even implicit Stmts should have proper ranges that include any
-  // non-implicit Stmts (fix Stmts created for lazy vars).
-  if (!S->isImplicit() &&
-      !rangeContainsLoc(Lexer::getCharSourceRangeFromSourceRange(
-          getSourceMgr(), S->getSourceRange())))
-    return false;
-  return !tryResolve(S);
-}
-
-bool CursorInfoResolver::walkToStmtPost(Stmt *S) {
-  if (isDone())
-    return false;
-  // FIXME: Even implicit Stmts should have proper ranges that include any
-  // non-implicit Stmts (fix Stmts created for lazy vars).
-  if (!S->isImplicit() && getSourceMgr().isBeforeInBuffer(LocToResolve,
-                                                          S->getStartLoc()))
-    return false;
-  return true;
-}
-
-bool CursorInfoResolver::visitDeclReference(ValueDecl *D,
-                                            CharSourceRange Range,
-                                            TypeDecl *CtorTyRef,
-                                            ExtensionDecl *ExtTyRef, Type T,
-                                            ReferenceMetaData Data) {
-  if (isDone())
-    return false;
-  if (Data.isImplicit)
-    return true;
-  return !tryResolve(D, CtorTyRef, ExtTyRef, Range.getStart(), /*IsRef=*/true, T);
-}
-
-bool CursorInfoResolver::walkToExprPre(Expr *E) {
-  if (!isDone()) {
-    if (auto SAE = dyn_cast<SelfApplyExpr>(E)) {
-      if (SAE->getFn()->getStartLoc() == LocToResolve) {
-        ContainerType = SAE->getBase()->getType();
-      }
-    } else if (auto ME = dyn_cast<MemberRefExpr>(E)) {
-      SourceLoc MemberLoc = ME->getNameLoc().getBaseNameLoc();
-      if (MemberLoc.isValid() && MemberLoc == LocToResolve) {
-        ContainerType = ME->getBase()->getType();
-      }
-    }
-    auto IsProperCursorLocation = E->getStartLoc() == LocToResolve;
-    // Handle cursor placement after `try` in ForceTry and OptionalTry Expr.
-    auto CheckLocation = [&IsProperCursorLocation, this](SourceLoc Loc) {
-      IsProperCursorLocation = Loc == LocToResolve || IsProperCursorLocation;
-    };
-    if (auto *FTE = dyn_cast<ForceTryExpr>(E)) {
-      CheckLocation(FTE->getExclaimLoc());
-    }
-    if (auto *OTE = dyn_cast<OptionalTryExpr>(E)) {
-      CheckLocation(OTE->getQuestionLoc());
-    }
-    // Keep track of trailing expressions.
-    if (!E->isImplicit() && IsProperCursorLocation)
-      TrailingExprStack.push_back(E);
-  }
-  return true;
-}
-
-bool CursorInfoResolver::walkToExprPost(Expr *E) {
-  if (isDone())
-    return false;
-  if (!TrailingExprStack.empty() && TrailingExprStack.back() == E) {
-    // We return the outtermost expression in the token info.
-    CursorInfo.setTrailingExpr(TrailingExprStack.front());
-    return false;
-  }
-  return true;
-}
-
-bool CursorInfoResolver::visitCallArgName(Identifier Name,
-                                          CharSourceRange Range,
-                                          ValueDecl *D) {
-  if (isDone())
-    return false;
-  bool Found = tryResolve(D, nullptr, nullptr, Range.getStart(), /*IsRef=*/true);
-  if (Found)
-    CursorInfo.IsKeywordArgument = true;
-  return !Found;
-}
-
-bool CursorInfoResolver::
-visitDeclarationArgumentName(Identifier Name, SourceLoc StartLoc, ValueDecl *D) {
-  if (isDone())
-    return false;
-  return !tryResolve(D, nullptr, nullptr, StartLoc, /*IsRef=*/false);
-}
-
-bool CursorInfoResolver::visitModuleReference(ModuleEntity Mod,
-                                              CharSourceRange Range) {
-  if (isDone())
-    return false;
-  if (Mod.isBuiltinModule())
-    return true; // Ignore.
-  return !tryResolve(Mod, Range.getStart());
-}
-
 SourceManager &NameMatcher::getSourceMgr() const {
   return SrcFile.getASTContext().SourceMgr;
-}
-
-bool CursorInfoResolver::rangeContainsLoc(SourceRange Range) const {
-  return getSourceMgr().rangeContainsTokenLoc(Range, LocToResolve);
-}
-
-bool CursorInfoResolver::rangeContainsLoc(CharSourceRange Range) const {
-  return Range.contains(LocToResolve);
 }
 
 std::vector<ResolvedLoc> NameMatcher::resolve(ArrayRef<UnresolvedLoc> Locs, ArrayRef<Token> Tokens) {
